pyproject.toml
[project]
name = "ib-report-formatter"
version = "1.0.0"
description = "Markdown to Professional IB-Style Word Document Converter"
readme = "AGENTS.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "DCM Team"}
]
keywords = ["markdown", "word", "docx", "investment-banking", "report", "converter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "python-docx>=0.8.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
full = [
    "matplotlib>=3.5.0",      # LaTeX equation rendering
    "charset-normalizer>=3.0", # Better encoding detection for Korean
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0,<1.18",
    "types-PyYAML>=6.0",
]

[project.scripts]
ib-report = "md_to_word:main"
md-format = "md_formatter:main"
roundtrip-audit = "roundtrip_audit:main"

[project.urls]
Homepage = "https://github.com/example/ib-report-formatter"
Documentation = "https://github.com/example/ib-report-formatter#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py38"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = [
    "cli_utils",
    "deep_md_cleaner",
    "ib_renderer",
    "md_formatter",
    "md_parser",
    "md_renderer",
    "md_to_word",
    "word_parser",
    "word_to_md",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
